+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Dynamically compute SOC values. [#11](https://github.com/paul121/farm_loocc/issues/11)
 - Delete estimates when assets are archived. [#10](https://github.com/paul121/farm_loocc/issues/10)
 - Add `deleteEstimate` function to `farm_loocc.estimate` service.